Stages + score record

+++ TryFrog/src/p4_group_8_repo/Turtle.java
@@ -6,7 +6,7 @@ public class Turtle extends Actor{
 	Image turtle1;
 	Image turtle2;
 	Image turtle3;
-	private int speed;
+	private double speed;
 	int i = 1;
 	boolean bool = true;
 	@Override
@@ -31,7 +31,7 @@ public class Turtle extends Actor{
 		if (getX() < -75 && speed<0)
 			setX(600);
 	}
-	public Turtle(int xpos, int ypos, int s, int w, int h) {
+	public Turtle(int xpos, int ypos, double s, int w, int h) {
 		turtle1 = new Image("file:src/p4_group_8_repo/TurtleAnimation1.png", w, h, true, true);
 		turtle2 = new Image("file:src/p4_group_8_repo/TurtleAnimation2.png", w, h, true, true);
 		turtle3 = new Image("file:src/p4_group_8_repo/TurtleAnimation3.png", w, h, true, true);
